feat: check containers before acc delete

+++ src/domain/useCase/deleteAccount.go
@@ -12,10 +12,21 @@ func DeleteAccount(
 	accQueryRepo repository.AccQueryRepo,
 	accCmdRepo repository.AccCmdRepo,
 	accountId valueObject.AccountId,
+	containerQueryRepo repository.ContainerQueryRepo,
 ) error {
 	_, err := accQueryRepo.GetById(accountId)
 	if err != nil {
 		return errors.New("AccountNotFound")
+	}
+
+	containers, err := containerQueryRepo.GetByAccId(accountId)
+	if err != nil {
+		log.Printf("GetContainersByAccIdError: %s", err)
+		return errors.New("GetContainersByAccIdInfraError")
+	}
+
+	if len(containers) > 0 {
+		return errors.New("AccountHasContainers")
 	}
 
 	err = accCmdRepo.Delete(accountId)
